rozplątanie zależności między plikami, ze względu na cykliczne importy po zmianach z CGS-20
