+++ api/app/blueprints/layers/routings.py
@@ -1,11 +1,17 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
+import uuid
+
+import jwt
+import psycopg2
 
 from flask import Blueprint, jsonify, request, current_app, make_response, Response, send_file
 from flasgger import swag_from
+
+from app.db.database import database
 from app.docs import path_by
-from app.db.general import token_required, cloud_decorator, layer_decorator
-from app.helpers.cloud import Cloud
+from app.db.general import token_required, layer_decorator
+from app.helpers.cloud import Cloud, cloud_decorator
 from app.helpers.layer import Layer
 from shapely.geometry import shape
 from psycopg2.sql import SQL, Identifier, Placeholder
@@ -21,7 +27,7 @@ from io import BytesIO
 import json
 
 mod_layers = Blueprint("layers", __name__)
-
+from app.blueprints.layers.tags import routings
 
 GDAL_TO_PG = {
     "String": "character varying",
