+++ api/app/blueprints/rdos/analysis/models.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-from app.db.general import BaseModel
+from app.db.base_model import BaseModel
 from peewee import TextField, DoesNotExist
 from playhouse.shortcuts import model_to_dict
 import os
