+++ api/app/db/base_model.py
@@ -1,0 +1,7 @@
+from peewee import Model
+
+from app.db.database import database
+
+class BaseModel(Model):
+    class Meta:
+        database = database
