+++ api/pytest.ini
@@ -5,6 +5,7 @@ markers =
     features: Features module
     layers: Layers module
     layerssettings:  Layers settings
+    layers_tags: Layers module - tags
     styles: Layer styles
     permissions: Permissions module
     export: Export module
